feat: allow secrets overriding by environment values, add values to local dotenv
