+++ apps/af-mvp/.env.local
@@ -9,3 +9,8 @@ USERS_API_ACCESS_KEY=secret
 NEXT_PUBLIC_LOGIN_SYSTEM_COGNITO_USER_POOL_ID=eu-north-1_DFVY9SQ6V
 NEXT_PUBLIC_LOGIN_SYSTEM_COGNITO_CLIENT_ID=hjbs3ag21jh24tpvn2a6q9nml
 NEXT_PUBLIC_LOGIN_SYSTEM_COGNITO_USER_POOL_ENDPOINT=https://cognito-idp.eu-north-1.amazonaws.com/eu-north-1_DFVY9SQ6V
+
+# Local secrets
+local_BACKEND_HASHGEN_KEY=secret
+local_BACKEND_SECRET_PRIVATE_KEY="-----BEGIN PRIVATE KEY-----\nMIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQCv0vArp/BwZxxj\nwg1PeICTCSgOpAWOxwtKT34xB0F+OV5eijYzYdIiMRvrFVncScB/ITf32SvgbQ0r\nhCerOv+6wDmHnK59imNcV4L4zpVDrA2+k9HMH3BH9KDcxjBzgNsKElifuvVOjY9M\nCOCAUVidguB8FWdJdVYxayE90cT90dbkvfEmZHp9IZ1+SMG6UNZ2EpcfdX5lIvag\nTrvJZmZUlAYiulQ+JRCFykn7zsdfhGfcfhDjydcoMuQnCNBmWPx8T4JUICcIjkY3\n4RNUYIYK+YdExzSOdk4R+5EVhB13WuETNU7T8GYuDecPS01h9GAtYHNbHu5nkMLo\nm0C1//ubAgMBAAECggEAB8rQ8xZbzRd8xh0NCvcbt6iBunhheKzwnqf8Pz8Dk40G\nVilI8qkEr5g+lQIau+0W12uI9/aLf8RFf6a5PVj9iXDIe+6hOxYdhjOorc82BUxE\nRax2cVX5W28HjBFKh2vwCTe77sRR59vVPHjuPjGcvEpXGZi3PaWyHrmRPRSlZtxJ\ncgjeks8dKUZ2YnU0vS8/Tw01Qm0TNF8EhdbWp/NhOWoEdjg6UmmtMzJjSqty+jhg\nmUr9szZUm/dyX1RjD/3lOdyzTZKXvs/TLu/qwfEldLer7ECckD52aS6AYqW7Fp+X\nN4EMDzb3YJo0AUjer5SSMKgzetK7Jc2ex64sJSQUwQKBgQDyJYDAEKhFLXhwHiaM\n2z63qFbRNCbqFj26UUSsFGk5vwqmJ//UDLUiNiUH5cjhlKr163zgVb6uQVSjee6u\nigi79elEP5iDSLyw1NoFe1Y0RB4v4Q9qXSYdsbJS1Q5iKxfz2CfNYfPOCLDMHDq4\nL2hZ1HrasQ1YcZB7ymQRrC/LwQKBgQC54hH+ejJ+WBx+QAsCM3cVkRehdjqPa9Rd\nxShs+grz1zkXVPdbbT1fiXUAxRcSYDbBmqob3kqiulpd3rOXiXdpIml9NTHO5tut\nB3EzUcFZaa67CZe5Mcuzx19cNLomnr5sc3uKrUIKMR+H2K2a7k3+v0eu34zwY32E\nILtR014OWwKBgQDxKtjoca6l93uUbvMGJvZ4BHkzMzXuvugLwRUcLdf/Y34/rmbl\n0yI0SKjpDLrdsSv0ce2VoNCgzQB3MOG/hEreqTUCfDRxzOfhVp0uCID63T0TgpvP\n26YyvOBWD0ycCT955Tb/HyerdIxhPVxeRRkxHqadWOlpxPGZ2SGBWHTiAQKBgQCb\nZrrwayynn/ZCPjx13JzkNObyn6sZmGZ09o1zfh+F+BrMvMYHc25p3rsop51ASZu9\nhBZPMipnnTBLZhHy0IQ6KwDSIpS/ixWAtMYYx3LD06y2xbBIra7FYBsDNTmja/d7\nJ5jTOUIKz0KzCZ0p96zRGw1Mz6w+lI6s1ciFAhaSHQKBgQCES8c96HL3WKCI46r4\n3SGIGLUYQlG3ay9KD0rmXyyyh6MucgfHiZyOY2yFND0zQDcIj83+AKBWeRMMnjVE\nMD1Cwm3GvcH9KPP/fmB9vE65AN0T1RQLw2DOOPhAmGuP/bB4G1eWy7PZ9AqVL8ZT\nDknnDoXCMVnRTPb/c39Pav8G7Q==\n-----END PRIVATE KEY-----"
+local_BACKEND_SECRET_PUBLIC_KEY="-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAr9LwK6fwcGccY8INT3iA\nkwkoDqQFjscLSk9+MQdBfjleXoo2M2HSIjEb6xVZ3EnAfyE399kr4G0NK4Qnqzr/\nusA5h5yufYpjXFeC+M6VQ6wNvpPRzB9wR/Sg3MYwc4DbChJYn7r1To2PTAjggFFY\nnYLgfBVnSXVWMWshPdHE/dHW5L3xJmR6fSGdfkjBulDWdhKXH3V+ZSL2oE67yWZm\nVJQGIrpUPiUQhcpJ+87HX4Rn3H4Q48nXKDLkJwjQZlj8fE+CVCAnCI5GN+ETVGCG\nCvmHRMc0jnZOEfuRFYQdd1rhEzVO0/BmLg3nD0tNYfRgLWBzWx7uZ5DC6JtAtf/7\nmwIDAQAB\n-----END PUBLIC KEY-----"
