+++ apps/af-mvp/src/lib/backend/services/aws/ParameterStore.ts
@@ -1,8 +1,6 @@
 import { GetParameterCommand, SSMClient } from '@aws-sdk/client-ssm';
-import { getRuntimeStage } from '@shared/lib/utils';
 
 const parameterStoreClient = new SSMClient({});
-const inMemoryCache: Record<string, string> = {};
 
 // @see: https://docs.aws.amazon.com/AWSJavaScriptSDK/v3/latest/clients/client-ssm/classes/getparametercommand.html
 export async function getSecretParameter(name: string) {
@@ -20,21 +18,4 @@ export async function getSecretParameter(name: string) {
     }
     throw error;
   }
-}
-
-export async function getStagedSecretParameter(name: string) {
-  const stage = getRuntimeStage();
-  return getSecretParameter(`${stage}_${name}`);
-}
-
-export async function getCachingSecretParameter(name: string) {
-  if (inMemoryCache[name]) return inMemoryCache[name];
-  const value = await getSecretParameter(name);
-  inMemoryCache[name] = value;
-  return value;
-}
-
-export async function getCachingStagedSecretParameter(name: string) {
-  const stage = getRuntimeStage();
-  return getCachingSecretParameter(`${stage}_${name}`);
 }
